Move the UserInfo into northbound security
The User Info class belongs with the northbound security API which is a factory for it
