+++ distribution/features/northbound-ogc-sensorthings-mqtt-feature/pom.xml
@@ -27,8 +27,8 @@
   <properties>
     <sensinact.version>0.0.2-SNAPSHOT</sensinact.version>
     <sensorthings.version>0.0.2-SNAPSHOT</sensorthings.version>
-    <moquette.version>0.16</moquette.version>
-    <netty.version>4.1.77.Final</netty.version>
+    <moquette.version>0.17</moquette.version>
+    <netty.version>4.1.93.Final</netty.version>
   </properties>
 
   <dependencies>
