+++ distribution/features/northbound-ogc-sensorthings-mqtt-feature/src/main/resources/northbound-ogc-sensorthings-mqtt-feature.json
@@ -2,17 +2,17 @@
   "id": "org.eclipse.sensinact.gateway.distribution.features:northbound-ogc-sensorthings-mqtt-feature:osgifeature:0.0.2-SNAPSHOT",
   "bundles": [
     { "id": "com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.16.1" },
-    { "id": "io.netty:netty-common:4.1.77.Final" },
-    { "id": "io.netty:netty-buffer:4.1.77.Final" },
-    { "id": "io.netty:netty-codec:4.1.77.Final" },
-    { "id": "io.netty:netty-codec-http:4.1.77.Final" },
-    { "id": "io.netty:netty-codec-mqtt:4.1.77.Final" },
-    { "id": "io.netty:netty-common:4.1.77.Final" },
-    { "id": "io.netty:netty-handler:4.1.77.Final" },
-    { "id": "io.netty:netty-resolver:4.1.77.Final" },
-    { "id": "io.netty:netty-transport:4.1.77.Final" },
-    { "id": "io.netty:netty-transport-classes-epoll:4.1.77.Final" },
-    { "id": "io.netty:netty-transport-native-unix-common:4.1.77.Final" },
+    { "id": "io.netty:netty-common:4.1.93.Final" },
+    { "id": "io.netty:netty-buffer:4.1.93.Final" },
+    { "id": "io.netty:netty-codec:4.1.93.Final" },
+    { "id": "io.netty:netty-codec-http:4.1.93.Final" },
+    { "id": "io.netty:netty-codec-mqtt:4.1.93.Final" },
+    { "id": "io.netty:netty-common:4.1.93.Final" },
+    { "id": "io.netty:netty-handler:4.1.93.Final" },
+    { "id": "io.netty:netty-resolver:4.1.93.Final" },
+    { "id": "io.netty:netty-transport:4.1.93.Final" },
+    { "id": "io.netty:netty-transport-classes-epoll:4.1.93.Final" },
+    { "id": "io.netty:netty-transport-native-unix-common:4.1.93.Final" },
     { "id": "org.eclipse.sensinact.gateway.northbound.sensorthings:dto:0.0.2-SNAPSHOT" },
     { "id": "org.eclipse.sensinact.gateway.northbound.sensorthings:mqtt:0.0.2-SNAPSHOT" }
 ],
