+++ distribution/features/northbound-oidc-authenticator-feature/pom.xml
@@ -14,7 +14,7 @@
   <dependencies>
     <dependency>
       <groupId>org.eclipse.sensinact.gateway.northbound.security</groupId>
-      <artifactId>authentication-api</artifactId>
+      <artifactId>security-api</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
