+++ northbound/rest/pom.xml
@@ -44,7 +44,7 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}.security</groupId>
-      <artifactId>authentication-api</artifactId>
+      <artifactId>security-api</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
