+++ northbound/rest/src/main/java/org/eclipse/sensinact/northbound/rest/impl/SensinactSessionProvider.java
@@ -14,8 +14,8 @@ package org.eclipse.sensinact.northbound.rest.impl;
 
 import java.security.Principal;
 
-import org.eclipse.sensinact.core.security.UserInfo;
 import org.eclipse.sensinact.northbound.rest.impl.AuthenticationFilter.UserInfoPrincipal;
+import org.eclipse.sensinact.northbound.security.api.UserInfo;
 import org.eclipse.sensinact.northbound.session.SensiNactSession;
 import org.eclipse.sensinact.northbound.session.SensiNactSessionManager;
