+++ northbound/security/security-api/src/main/java/org/eclipse/sensinact/northbound/security/api/Authenticator.java
@@ -12,8 +12,6 @@
 **********************************************************************/
 package org.eclipse.sensinact.northbound.security.api;
 
-import org.eclipse.sensinact.core.security.UserInfo;
-
 public interface Authenticator {
 
     public UserInfo authenticate(String user, String credential);
