+++ northbound/security/security-api/src/main/java/org/eclipse/sensinact/northbound/security/api/UserInfo.java
@@ -10,7 +10,7 @@
 * Contributors:
 *   Kentyou - initial implementation
 **********************************************************************/
-package org.eclipse.sensinact.core.security;
+package org.eclipse.sensinact.northbound.security.api;
 
 public interface UserInfo {
