+++ northbound/websocket/pom.xml
@@ -45,7 +45,7 @@
     </dependency>
     <dependency>
       <groupId>org.eclipse.sensinact.gateway.northbound.security</groupId>
-      <artifactId>authentication-api</artifactId>
+      <artifactId>security-api</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
