+++ northbound/websocket/src/main/java/org/eclipse/sensinact/northbound/ws/impl/WebSocketCreator.java
@@ -19,8 +19,8 @@ import java.util.List;
 import org.eclipse.jetty.websocket.server.JettyServerUpgradeRequest;
 import org.eclipse.jetty.websocket.server.JettyServerUpgradeResponse;
 import org.eclipse.jetty.websocket.server.JettyWebSocketCreator;
-import org.eclipse.sensinact.core.security.UserInfo;
 import org.eclipse.sensinact.northbound.query.api.IQueryHandler;
+import org.eclipse.sensinact.northbound.security.api.UserInfo;
 import org.eclipse.sensinact.northbound.session.SensiNactSessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
